Expand docker upgrade test, sunset E2E upgrade test
