+++ tests/docker/resources/daemonset.yaml
@@ -1,0 +1,18 @@
+apiVersion: apps/v1
+kind: DaemonSet
+metadata:
+  name: test-daemonset
+spec:
+  selector:
+    matchLabels:
+      k8s-app: test-daemonset
+  template:
+    metadata:
+      labels:
+        k8s-app: test-daemonset
+    spec:
+      containers:
+        - name: webserver
+          image: rancher/mirrored-library-nginx:1.29.1-alpine
+          ports:
+          - containerPort: 80
